rename single main files to index.tsx
